CQID: WIND00155220
[
Using KGDBoE with the smc91x ethernet device is unusable since it
locks up frequently no longer sending or receiving data. This patch
does several things to prevent this.

Several functions have had their locking updated to match what is
is seen with the smc911x driver and also to fix perceived errors
in the locking. The receive function has been modified to pull
all it can pull from the receive fifo allowing for better data
flow and memory usage. The watchdog timeout has been corrected
to match the comment and what is seen in the smc911x. A mdelay
has been added to the poll routine as it has been found that
without the delay reading the Interrupt Register at a high
frequency causes issues (basically the register reports the
same value even though it should in fact report a receive
interrupt event).

Basically I have spent many hours looking at this problem and
was slowly able to rule out QEMU, KGDB, GDB and others and
put the blame on the ethernet driver and the device. This problem
is definitely an issue on both QEMU and real hardware yet it is
much more evident on QEMU. I was not able to determine why the
interrupt status register gets stuck but it does and there is
nothing that can be done to either detect that it is stuck or
reset things to unstick it. So basically the solution of not
pounding on it is the best I can come up with.
]

removed: bsp/arm_versatile_926ejs/arm_versatile_926ejs-temporary-smc-patch-for-netpol.patch

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>