Split out R9721 CPU reference from performanc counter driver patch
CQID:WIND00155646

The patch we added to the oprofile feature had reference to an
undefined CPU type (R9721). This CPU type is only introduced
after the NEC EMMA3P patches are applied. So basically we remove
them from the patch and introduce an emma3p specific patch to
re-add them.

In the end this is a nicer approach as the original patch is
in better shape for upstream submission.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>