kexec: all HW not initialized correctly
CQID: WIND00146141

[
After running a Kexec'd kernel not all registers are left
in the same state as the bootloader leaves them.  Unfortunately
the driver assumes that the interrupt error mask register is
zeroed by the bootloader, which on this board is U-Boot.
This is clearly not the case when running the fastboot kernel.
Zero out the interrupt error register on the host controller.
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>