Clock-api testing failed on raza_xlr_atx_64_be-std-glibc_std
CQID:WIND00155449

[
The original design is that vCPU0 (the first boot up thread)
use PIC's timer7 as its timer interrupt source (clock event)
and other 30 vCPUs use mips compare/count as its timer
interrupt source.

This design is not symmetric and it make the clock api and
HRT(High Resolution Timers) difficult to implement.

So use the symmetric implementation (vCPU0 also use the
compare/count as its clock event). And this implementation make the
HRT feature of CGL easy to support.

The new design is all vcpus use the compare/count clock event and
the clock api had been implemented in cevt-r4k.c. So the original
added code is useless. And we need to add the clock api smp support.

The original implementation of PIC timer7 clock event is useless.
So we need to remove it.

[PATCH] make the cpu0 use the compare and count clock event
[PATCH] enable internal clock interrupt of vcpu0
[PATCH] remove the unusable code of clock api
[PATCH] remove pic timer7 clock event implementation
[PATCH] clock event and clock api kconfig

Signed-off-by: Jack Tan <jack.tan@windriver.com>
]

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>