smc91x driver reports some RX warnings/errors
CQID: WIND00156032

[
This was only added to the original fix for QEMU/KGDBoE issue
we observed with the arm_versatile as a mitigation since
we were adding a delay and thus this will help ensure the
RX queue was being kept empty as efficiently as possible.
However this has been observed to cause issues so I am
removing the mitigation and returning things to how they
were.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>