atca6101: fix kexec/kdump
CQID: WIND00141946

[
Rebooting into a crash kernel on emer_atca6101 failed mounting the
filesystem. This was due to two things: the IOMMU being turned off for
failing a test on the amount of memory available to the kernel and
interrupts from the ethernet device not being seen by the CPU.

The check for the amount of memory for the IOMMU seems to be arbitrary:
the IOMMU only needs a maximum of 2MB (from comment in code) but still
looks for >1GB of RAM being present before turning itself on. The crash
kernel tries to be as small as possible, ie. 128MB. The first patch
circumvents the check for the 6101.

The root cause of the second problem has not been fully determined, only
that level-trigger interrupts do not function properly on the crash
kernel,
possibly due to the hardware being left in an inconsistent state
following crashes. Edge-trigger interrupts function normally, the serial
port being responsive. A workaround has been added for a 6101/crash
kernel combination, which runs the front panel ethernet device in poll
mode.
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>