kconfig sun_niagara2_bsp disable unnecessary serial drivers
CQID: WIND00156134

On some Netra SUN4V targets such as the T-5220 or T-5440, the serial console
could be provided through either the sunsu (device emulation) or sunhv
(hypervisor) serial drivers.

Prefer the sunhv driver or the sunsu driver, as the sunsu driver lacks
the serial console poll handlers necessary for kgdboc to work.

Also, disable the sunsab driver as it's not needed for the Netra SUN4V targets.

Signed-off-by: Boban Petrovic <boban.petrovic@windriver.com>
Signed-off-by: Vu Tran <vu.tran@windriver.com>
Signed-off-by: Hong H. Pham <hong.pham@windriver.com>