emer_atcaf101: fix locking to support kgdboe
CQID:WIND00156189

[
According to Jasson's comments, while the BSP driver had a spin_unlock_irq()
somewhere, and if the board continued on and will hit the exception a
second time. This results in the following panic:
===
root@ATCAF101-kyang:/> ls
Kernel panic - not syncing: Recursive entry to debugger
Rebooting in 180 seconds..
===

For F101, it's safe to replace spin_lock_irq/unlock
with spin_lock_irqsave/restore on eth's poll function.
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>