sparc64: Fix msgrcv() hang
CQID: WIND00156176

[
The glibc-2.8 SPARC64 SysV IPC msgrcv() handler uses the old API, which
requires passing an ipc_kludge structure as the fifth argument.  The
Linux kernel's SPARC64 sys_ipc msgrcv handler does not check for the backward
compatibility hack, so a bogus "msgflag" parameter gets passed down to
the real SysV IPC handler.  This causes msgrcv() to hang.

The fix is to implement the msgrcv backwards compatibility hack, despite
the comment in sys_ipc() that there's no need for this hack.

Arguably the fix should be in glibc as it's using a deprecated syscall
API, but it's way easier to put the fix in the kernel.
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>