ftrace: disable by default for mips AND enforce it
CQID: WIND00156343

It was discovered that ftrace for MIPS and modules don't
mix. The fix is involved and not something we can get
done for 3.0.

We could have done something in Kconfig and made the
two mutually exclusive, but that involves more risk.
So the solution is to prevent our config fragments
from enabling ftrace on MIPS, unless you do an
explicit override.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>