header cleanup: change some of the patch headers for autocommit
CQID: 145801

The way that guilt processes the patches, leaves some of them
with a commit author of autocommit@autocommit, and if there
is no patch header at all, then it creates an auto message
as well.

This fixes up the first ~25 RT patches to have decent headers
w/o any autos, and the trailing RT patches which were WR creations
as well.  Plus some of the ones for wrnote etc.

There are no functional changes in this patch.  Since you have
to do a "guilt refresh" after doing a "guilt header -e" you
get some patches like the omap one having refresed line numbers
as well as the header change.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>