perfmon3: remove optional feature
CQID: WIND00145801

perfmon3 was a stretch goal, but it not supported
in 3.0 and indeed won't compile. We need to remove
this development artifact else we'll end up supporting
it. This was not used and carries no risk.

upstream, perfcounters it competing with perfmon for
inclusion, so it's best to wait and see what gets
merged.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>