header cleanup: change misc std kernel patch headers for autocommit
CQID: 145801

The way that guilt processes the patches, leaves some of them
with a commit author of autocommit@autocommit, and if there
is no patch header at all, then it creates an auto message
as well.

These patches have been updated by manually adding a From and
a Subject as required.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>