header cleanup: change preempt_irq patch headers for autocommit
CQID: 145801

The way that guilt processes the patches, leaves some of them
with a commit author of autocommit@autocommit, and if there
is no patch header at all, then it creates an auto message
as well.

This fixes up the first ~25 RT patches to have decent headers
w/o any autos, and the trailing RT patches which were WR creations
as well.  Plus some of the ones for wrnote etc.

There are no functional changes in this patch.  Since you have
to do a "guilt refresh" after doing a "guilt header -e" you
get some patches one having refresed line numbers as well as
the header change.

I've audited the git-diff for this change, the only chunk that
isn't clearly just a line shift is this:

=========================================================
        }

 -      __get_cpu_var(netdev_rx_stat).total++;
 +      per_cpu(netdev_rx_stat, raw_smp_processor_id()).total++;

+       trace_net_dev_receive(skb);
        skb_reset_network_header(skb);
-       skb_reset_transport_header(skb);
=========================================================

...but it still is just a context update; the indented [+-]
of the original patch remain unchanged.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>