sparc msgrcv: refresh due to patch relocation
CQID: WIND00156176

patch was placed on end of BSP, but relocation puts it pre-ltt.

Remove ltt context from sparc patch and put new sparc content
into ltt context.

Note that this is the 1st time that the patch:

	 sparc64-lttng-hooks.patch

has been refreshed or changed, and as such, it gets reduced down
to 3 lines of context on either side of the deltas, vs. the original
eight that the original patch from Hong had.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>