context mods: move to 18, sparc Kconfig clash
syscall wrapper change of 28387e82363 bumps into some of the
existing sparc and sparc/ltt patches.

Also with trivial context mods, are an EDAC patch and a
grsec patch.

Note that one of the sparc patches gets condensed down from 8 lines
of context to the standard of three, which makes the change look
artifically big.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>