cgroup: fix kernel oops in cgroup scheduler
CQID: WIND00156551

Analysis:  There are 37 instances of the string "struct nlattr *tb" in
net/sched.

36 of them have a size of the form [FOO_MAX + 1].

One of them does not.

When you use this one, the kernel sometimes crashes in really weird
ways.  The characteristic behavior is that a local variable which was
living in a register gets zeroed out.  How?  Simple!  The register is
callee-saved.  The callee-saved registers are stored on the stack.
The last register saved is saved lowest on the stack -- immediately
above this array.

nla_parse_nested calls
        memset(tb, 0, sizeof (struct nlattr *) * (maxtype + 1));

Signed-off-by: Peter Seebach <peter.seebach@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>