oprofile: backtrace even if PMR registers not implemented
CQID: 157093

[
If the PPC variant does not have a Performance Monitor Register
interface, then the oprofile driver code uses the Real Time Clock
to generate interrupts.  However, the missing PMR detection code
caused a bypass of setting the backtrace callback, disabling the
backtrace code.

This fix unconditionally sets the backtrace callback prior
to testing for a PMR interface.  All of the other callbacks are
reset in oprofile_timer_init() which sets the RTC as the
interrupt generator when PMR detection fails.
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>