rcupreempt: use raw lock for schedlock
CQID: 150713

The original RT patch in 26 named call_rcu_sched.patch is
upstream in 27-rc1 as commit 4446a36, but since it is the
basic (i.e. non-RT aware version) the locks it messes with
are not raw.  This adds back in the delta between the RT
version of the patch and the upstream commit.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>