speculative get page: merge deltas vs e286781d
CQID: 150713

At a glance it would appear that the content of the 26-RT patch
2.6.21-rc6-lockless6-speculative-get-page.patch is present in 27rc1
as commit e286781d.  However it is not quite 100% the same.  The
obvious difference is the use of count vs NoNewRefs for locking,
but this delta here is also missing; since it is specific to RT.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>