stop_cpu: add stop_cpu-completion.patch
CQID: 150713

Currently, stop_cpu() calls ack_state(), which when all is done
(i.e. state == STOPMACHINE_EXIT), launches the completion to tie
up loose ends.  However, we are still running with irqs disabled
from what stop_cpu() set up earlier, and so we get a warning.

Since we know we are out the bottom already and in final EXIT
stage, we can add a patch to enable irqs around the completion.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>