oom: allow disabling of OOM protection & normalize OOM protection size
CQID: WIND00162420

[
Allow user space application to reduce its OOM protection bytes
so that it is less likely to be exempt from OOM death.

Signed-off-by: Harry Ciao <qingtao.cao@windriver.com>

The total_vm is in the unit of PAGE_SIZE, so a process's OOM protection
bytes should be translated into PAGE_SIZE as well when comparing with
its total_vm.

Signed-off-by: Harry Ciao <qingtao.cao@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>