oprofile: fix x86 oprofile as module issues
CQID: WIND00164993

[
When oprofile is loaded as a module, undefined symbols
are reported.  This is fixed by use of a structure to
hold the system call return addresses which are used
by oprofile backtrace during syscall unwinding (0001-....patch).
The structure address values are set by code n the x86/kernel
directory and this code is always built-in when oprofile is enabled.
When oprofile is built as a module, the syscall return addresses
are available through this structure.  The syscall return
addresses are not simply EXPORT'd for security reasons.

The system call functions do not properly construct
the special oprofile backtrace frame when oprofile is built
as a module, because the conditional compilation is based
on CONFIG_OPROFILE and when a module only CONFIG_OPROFILE_MODULE
is set.  Instead of providing a compound conditional on both
CONFIG_OPROFILE and CONFIG_OPROFILE_MODULE, a new config
option is provided, CONFIG_OPROFILE_SYSCALL_TRAVERSE (0002...patch)
This has the advantages
        1) of enabling the changes in the system call frame
        layout for oprofile in both built-in and oprofile-as-module,
and
        2) allowing the customer to disable the special frame layout
        when using oprofile, if for example, oprofile is not
        being used with workbench and the customers wants no
        additional instructions in the system call to support
        workbench profiling.
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>