emer_atca6101: ensure that all CGL patches are on ECGL branches
The emer_atca6101-ecgl board description wasn't including the
CGL variant, this left the possibility of CGL patches not being
applied to the ECGL based BSPs. The fix is simple, we create
a common feature description for board specific CGL patches
and include it from both CGL and ECGL based boards.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>