cavium/ethernet: fix IPv6 ping
CQID: WIND00172371

[
Change the cavium ethernet driver so that it correctly calculates the
skb->tail to be either an absolute address or offset-based depending on
how the kernel is configured.  Currently it is only absolute based and
this causes problems with utilities that use the skb->tail address
for calculations.  Specifically, this was seen in testing as IPv6 pings
(via ICMP6) were failing due to neighbour association requests
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>