MIPS: system snapshot support
1/1 [
Author: Ralf Baechle
Email: ralf.baechle@windriver.com
Subject: MIPS: Enable kernel snapshots.
Date: Thu, 16 Jul 2009 19:49:27 +0100

This allows taking a snapshot of the state of a running system.  The
snapshot file will be a standard core file so all common tools can be
used to analyze it.  To take a snapshot just echo the filename of the
snapshot to be created into the control file like
"echo /root/snapshot > /sys/ksnap/ksnap".

Depending on the amount of memory in the system and filesystem performance
taking a snapshot may be a slow process; to allow a consistent snapshot
being taken the system will be mostly frozen during the time of the
snapshot.

The implementation is restricted to non-NUMA, non-HIGHMEM systems.

Signed-off-by: Ralf Baechle <ralf.baechle@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>