kernel: booting issues with cavium
CQID: WIND00175646

1/2 [
Author: Greg Moffatt
Email: greg.moffatt@windriver.com
Subject: mips/octeon: double unlock in memory setup
Date: Mon, 10 Aug 2009 11:28:08 -0400

For cavium the bootmem semaphore is unlocked twice during boot.
This fix changes it so that the semaphore is only unlocked once
in all scenarios.

Signed-off-by: Greg Moffatt <greg.moffatt@windriver.com>
]

2/2 [
Author: Greg Moffatt
Email: greg.moffatt@windriver.com
Subject: mips/octeon: booting cavium with 4GB+ RAM fails
Date: Mon, 10 Aug 2009 11:28:09 -0400

When booting a cavium board that has 4GB or more of ram the board will
die silently and the boot fail.  The root cause is due to the way the
bootloader passes the DRAM size to the kernel.  The bootloader passes
the DRAM size in MB as a uint32_t.  When the kernel starts up, it
converts it to bytes by bitshifting 20.  Because the bootloader is uint32_t
bitshifting 20 causes the result to be zero.  The solution is to cast the
bootloader DRAM size into a uint64_t before bitshifting.

Signed-off-by: Greg Moffatt <greg.moffatt@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>