rmi: EDAC fixes
CQID: WIND00175639 WIND00175640

1/1 [
Author: Ralf Baechle
Email: ralf.baechle@windriver.com
Subject: rmi: EDAC fixes
Date: Tue, 11 Aug 2009 15:47:16 +0100

In violation of all Linux conventions the macros to access on-chip register
on RMI take a register number, not a register address.  So divide all the
offsets from the base register address by 4.  We can miraculously seem to
get away with this bug most of the time if the rmi-edac-dram driver is
compiled into the kernel while the bug manifests itself consistently if the
driver is a module.

Signed-off-by: Ralf Baechle <ralf.baechle@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>