drivers/net: netpoll fixes
1/1 [
Author: DDD
Email: Dongdong.deng@windriver.com
Subject: net: Add init_dummy_netdev() and fix EMAC driver using it
Date: Sun, 27 Sep 2009 11:13:50 +0800

commit 937f1ba56b4be37d9e2ad77412f95048662058d2 upstream

This adds an init_dummy_netdev() function that gets a network device
structure (allocation and lifetime entirely under caller's control) and
initialize the minimum amount of fields so it can be used to schedule
NAPI polls without registering a full blown interface. This is to be
used by drivers that need to tie several hardware interfaces to a single
NAPI poll scheduler due to HW limitations.

It also updates the ibm_newemac driver to use that, this fixing the
oops on 2.6.29 due to passing NULL as "dev" to netif_napi_add()

Symbol is exported GPL only a I don't think we want binary drivers doing
that sort of acrobatics (if we want them at all).

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Tested-by: Geert Uytterhoeven <Geert.Uytterhoeven@sonycom.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Integrated-by: Dongdong Deng <dongdong.deng@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>