tipc: add room for DPA header in skb
CQID: WIND00186612

1/1 [
Author: Greg Moffatt
Email: greg.moffatt@windriver.com
Subject: tipc: add room for DPA header in skb
Date: Fri, 16 Oct 2009 15:05:27 -0400

TIPC allocates the skb's directly and passes those down for sending
the packet.  When TIPC allocates the skb it reserves some space in
the skb buffer for some header data.

On the P4080 target the DPA driver is used for packet transmission.
When the DPA driver is invoked it also attempts to allocate some space
in the passed-in skb for it's own header data (64 bytes).
Unfortunately the skb buffer allocated by TIPC is not large enough
to account for the TIPC data, TIPC header, and the DPA header.
As a result, the skb underflows and causes a kernel panic.

The solution is to use the LL_RESERVED_SPACE macro in the TIPC
ethernet send_msg() function.  Here the skb is checked to ensure
there's enough space for the driver to transmit the packet and
if not, the packets headroom is increased.

Signed-off-by: Greg Moffatt <greg.moffatt@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>