kgdb: update jiffies fix for non GENERIC_CLOCKEVENT boards
CQID: WIND00186904

The fix for updating jiffies when resuming from a break didn't include
a stub function for boards without GENERIC_CLOCKEVENTS. The mindspeed
router is one such board. The fix is simple, copy the stub implementation
for boards that don't support CONFIG_TICK_ONESHOT, and use it for this
case.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>