vbi: kill off confusing use of PPC32 within x86
1/2 [
Author: Paul Gortmaker
Email: paul.gortmaker@windriver.com
Subject: vbi: kill off confusing use of PPC32 within x86
Date: Tue, 27 Oct 2009 12:59:58 -0400

There is a use of "PPC32" in the vbi code with respect to the
setting of "CPU".  Unfortunately it is confusing, because it is
only set for x86 (?!) and it kind of gets confused with the
kernel's already existing CONFIG_PPC32.  Fortunately it can be
removed completely.  The existing trivial use case is shown below:

Current define:

  If CONFIG_X86=y
	PPC32 set to -1

  If CONFIG_PPC=y
	PPC never set

Current user (only the one!):
  #if ((CPU == PPC604)   || (CPU == PPC85XX) || \
      (CPU == PPCE500MC) || (CPU == PPC970))
  ...
  #elif (CPU != PPC32)
  ...
  #endif

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
]

2/2 [
Author: Paul Gortmaker
Email: paul.gortmaker@windriver.com
Subject: vbi ppc: clean up syscall file
Date: Tue, 27 Oct 2009 00:30:25 -0400

This is a cosmetic update, but the important thing that it does
is reinforce the need/ability to coalesce all the syscall defines
into a simple table form.  With all the extra fluff gone, it is
clear that most of them are the same.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>