8250: cleanup UART patching ; remove BSP specific patch
The standard branch had a bunch of WRHV specific changes to the
8250 driver, and then the fsl_8572 branch undid 95% of them.

We want the 8250 changes to be non BSP specific.  This moves
the 8572 patch back into standard, and hence the undo part
makes the existing "standard" patch smaller as the "do" and
"undo" chunks annihilate each other.

The 8250 code running for 8572 remains unchanged with this
reorder/relocation.  The x86 version does see the undo chunks
for the 1st time, but supermicro BSP was tested and functional.

Tested-by: Jim Somerville <Jim.Somerville@windriver.com>
Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>