x86/paravirt: make arch_flush_lazy_mmu/cpu disable preemption
CQID: 188072

1/1 [
Author: Jeremy Fitzhardinge
Email: jeremy@goop.org
Subject: x86/paravirt: make arch_flush_lazy_mmu/cpu disable preemption
Date: Mon, 2 Nov 2009 20:42:52 -0800

Impact: avoid access to percpu vars in preempible context

They are intended to be used whenever there's the possibility
that there's some stale state which is going to be overwritten
with a queued update, or to force a state change when we may be
in lazy mode.  Either way, we could end up calling it with
preemption enabled, so wrap the functions in their own little
preempt-disable section so they can be safely called in any
context (though preemption should never be enabled if we're actually
in a lazy state).

(Move out of line to avoid #include dependencies.)

Signed-off-by: Jeremy Fitzhardinge <jeremy.fitzhardinge@citrix.com>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
[Backport upstream commit d85cf93d. solved patch conflict]
Signed-off-by: Liang Li <liang.li@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>