Fix WIND00190244, WIND00190245 - cavium I2C can't detect eeprom
CQID: 190244

Summary: Correct I2C adapter class, allow I2C to build properly as a module
Tech Review: Mark Asselstine
Gatekeeper: Paul Gortmaker

Comments (indicate scope for each "y" above):
---------------------------------------------

This replaces an entire patch file in the 3.0.1 and 3.0.2 Cavium
Octeon async BSPs, but the actual change consists of two minor fixes.

The first change is in drivers/i2c/busses/i2c-octeon.c:

This change was present in the cav_ebt5800 BSP, but was omitted from
the cav_octeon BSP due to uncertainty as to whether the cav_ebt5800
BSP or Cavium's SDK 1.9 patches should take precedence.

The second change is in arch/mips/cavium-octeon/i2c-octeon_twsi.c:

This change resolves a problem with use of Octeon TWSI_RAW driver when
I2C_OCTEON_TWSI is compiled as a module. This option is selected in
the configuration when transfers in excess of 8 bytes are needed, such
as when attempting to dump EEPROM contents. (The I2C controller
hardware apparently can only deal with transfers of 8 bytes or less.)

Signed-off-by: Phil Staub <Phil.Staub@windriver.com>