MILS cleanup
CQID: WIND00187545

[

1/2 MILS: [cleanup] remove hardcoded bootLine offset

Instead of having a hardcoded value to define the offset
in which the bootLine lives.  Utilize the power of the compiler
to calculate respective values.

Signed-off-by: Jeremy McNicoll <jeremy.mcnicoll@windriver.com>

2/2 MILS: [cleanup] refactor SHIM syscall changes

The original solution did not make it easy to update the header
files with the latest VBI changes.  This should make life a
little bit easier.

Signed-off-by: Jeremy McNicoll <jeremy.mcnicoll@windriver.com>

]

I've also added some whitespace cleanups to the MILS patch.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>