MILS: rebase patch queue to place shim as last patch
CQID: WIND00187545

The previous direct export from kernel-multicore also had
a change to move/keep the shim patch at the end of the
patch list.  This commit contains the rebase of the patches
which were impacted by having it applied 44 deep vs at
the end of the queue.

You can now run the scc file w/o application failure, and the
tree created by running this queue is equiv. to the tree
out there now that has the shim patch 44 deep.

One of the shim cleanups has been folded in as well.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>