Stop using fixmap for hypervisor global regions
CQID: WIND00193612

1/1 [
Author: Jim Somerville
Email: Jim.Somerville@windriver.com
Subject: x68: stop using fixmap to map in hypervisor regions
Date: Wed, 9 Dec 2009 12:47:14 -0800

Instead use reservetop on the bootargs to set aside memory
space at the top of the address map.  Then map in the
hypervisor global regions directly.  This effectively
results in the exact same mapping as before, except that
we won't be interfering with a customer's potential use
of the fixmap mechanism.  This also removes one of our
hypervisor hooks, which is a good thing.

Signed-off-by: Zwane Mwaikambo <Zwane.Mwaikambo@windriver.com>
Signed-off-by: Jim Somerville <Jim.Somerville@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>