preempt_rt: rebase onto the end of the standard branch
RT patches were originally applied at the earliest opportunity
so that there would be no impact on the ~450 RT patches from
our patches.  But that prevents sustaining from using a common
base (i.e. end of standard) for staging CVE fixes etc. prior
to merging to all BSPs.

By rebasing RT to the end of standard, we take the one time
hit of going over the 450 patches, but sustaining will have
an easier time with things.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>