p2020: use mainline commit that is required for SMP
The P2020 BSP was using the file:

    0001-powerpc-85xx-Add-SMP-support-to-MPC8572DS.patch

which was incorrectly listed as having the following legacy:

   commit id: 6b347c3e111ec8d62429556f79852677328c45ea in mainline.

This commit number never was in mainline.  It was specific to
Kumar's public branch.

Use the mainline equivalent, which is 8bd3947afda14625bff8c067b76.
We've already added this commit to the standard branch.

The mainline imports an 8572 defconfig, which doesn't matter
in our use case.  The core SMP change we need is there.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>