85xx smp: track source of "powerpc/mm: Split mmu_context handling"
We've got a variant of the upstream commit ID 5e696617c425eb97bd943d
in tree for two BSPs, but the delta between upstream and this
commit prevents direct import into standard for complete sharing.

List one of the upstream sources for assistance in further
assessment for a push-down into standard.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>