kexec: disable PAX_UDEREF just before jumping
CQID: WIND00200299

1/1 [
Author: Greg Moffatt
Email: greg.moffatt@windriver.com
Subject: kexec: disable PAX_UDEREF just before jumping
Date: Fri, 12 Feb 2010 09:01:04 -0800

For CGL 32-bit kernels on the x86, PAX enables the configuration
option PAX_MEMORY_UDEREF.  By doing this any attempts to kexec from
this kernel will fail due to a GPF during the last moments before
jumping to the new kernel.  The GPF is caused by kexec.  The last
step before jumping to the new kernel is to reset all the segment
registers and invalidate the GDT.  With PAX_MEMORY_UDEREF enabled
this triggers a GPF on the kernel.

To rectify this the underlying functionality of PAX_MEMORY_UDEREF is
disabled just before jumping to the new kernel.  The PAX_MEMORY_UDEREF
can be dynamically disabled on the kernel commandline (see the
pax_nouderef commandline option).  To disable UDEREF kexec emulates
the functionality of the commandline handler just before the step
where kexec invalidates the GDT and jumps to the new kernel.

Signed-off-by: Greg Moffatt <greg.moffatt@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>