build: serialize image targets
1/2 [
Author: Bruce Ashfield
Email: bruce.ashfield@windriver.com
Subject: arm: serialize build targets
Date: Tue, 10 Jun 2008 21:58:35 -0400

The image targets need to trigger after we've build all their
dependencies. To do this we add .NOTPARALLEL and  ensure that
the image targets are not built in parallel.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>
]

2/2 [
Author: Bruce Ashfield
Email: bruce.ashfield@windriver.com
Subject: powerpc: serialize image targets
Date: Tue, 10 Jun 2008 21:58:45 -0400

The image targets need to trigger after we've build all their
dependencies. To do this we add .NOTPARALLEL and  ensure that
the image targets are not built in parallel.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>