octeon: Cavium Octeon specific kernel TLS optimization from SDK 1.9
1/2 [
Author: Yang Shi
Email: yang.shi@windriver.com
Subject: MIPS: Octeon: Add Octeon specific TLS thread pointer access optimization support
Date: Mon, 1 Mar 2010 22:07:50 -0800

For MIPS, normally the TLS thread pointer is accessed by the
userspace program executing a "rdhwr" from register $29. This
register doesn't exist, so the kernel emulates the instruction
assigning the thread pointer to the value register. This option
supplies an alternate, faster access to the thread pointer. A
side effect of this option is that the highest 8 bytes of CVMSEG
is used by the kernel to save and restore the thread pointer during
the TLB fault handlers. This CVMSEG address isn't available to user
applications.

Signed-off-by: Yang Shi <yang.shi@windriver.com>
]

2/2 [
Author: Yang Shi
Email: yang.shi@windriver.com
Subject: MIPS: Octeon: Replace slower references to the thread pointer with fast accesses
Date: Mon, 1 Mar 2010 22:07:51 -0800

With this optimization, the kernel can dynamically replace slower
references to the thread pointer with fast accesses. This involves
replacing userspace instructions at runtime, so it may not work with
all programs. It is advised to use a toolchain that creates code for
FAST_ACCESS_TO_THREAD_POINTER instead of this option. If you have
code compiled with a Cavium compiler prior to release 1.5, or are
using a non Cavium compiler, this option may allow you to receive
most of the benefit of direct access to the thread pointer. It may
also cause programs to fail.

Instruction replacement is disabled on boot. It can be controlled by
writing a mode to /sys/module/traps/parameters/thread_pointer_mode.
The supported modes are:

	0 - Use the normal kernel emulation without any changes.
	1 - Replace emulated instructions with direct accesses to the thread
		register.
	2 - Replace emulated instructions and log the replacement PC.
	3 - Replace emulated instructions with break instructions. This will
		cause programs to fail, but makes it easy to stop gdb on the
		instruction.

Signed-off-by: Yang Shi <yang.shi@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>