ppc: purge crap whitespace changes from patch
We'll need to ensure that this is 100% safe for deployment
on all non-multicore kernels -- having random whitespace
changes just detracts from this effort.

This should be a no-op code-wise.  You pretty much have to
dump this commit to a file and use ":se list" in vi to
see the untanglement.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>