CF broadcom_bcm91250a_64_be/le 4.0kernel patches
1/2 [
Author: Jack Tan
Email: jack.tan@windriver.com
Subject: RTC: SMBus support for the M41T80 (v2)
Date: Fri, 19 Mar 2010 16:23:19 +0800

The BCM1250A SOC which is used on the SWARM board utilising an M41T81
chip only supports pure I2C in the raw bit-banged mode.  Nobody sane
really wants to use it unless absolutely necessary and the M41T80, etc.
chips work just fine with an SMBus adapter which is what the standard mode
of operation of the BCM1250A.  The only drawback of byte accesses with the
M41T80 is the chip only latches clock data registers for the duration of
an I2C transaction which works fine with a block transfers, but not
byte-wise accesses.

The driver currently requires an I2C adapter providing both SMBus and raw
I2C access.  This is a set of changes to make it work with any SMBus
adapter providing at least read byte and write byte protocols.
Additionally, if a given SMBus adapter supports I2C block read and/or
write protocols (a common extension beyond the SMBus spec), they are used
as well.  The problem of unlatched clock data if SMBus byte transactions
are used is resolved in the standard way.  For raw I2C controllers this
functionality is provided by the I2C core as SMBus emulation in a
transparent way.

Signed-off-by: Maciej W. Rozycki <macro@linux-mips.org>

[ Original patch taken from this linux-mips kernel mailing list archive:
http://www.linux-mips.org/archives/linux-mips/2008-05/msg00242.html ]

Integrated-by: Weiwei Wang <weiwei.wang@windriver.com>
Integrated-by: Jack Tan <jack.tan@windriver.com>
]

2/2 [
Author: Jack Tan
Email: jack.tan@windriver.com
Subject: add the oprofile support of bcm1250
Date: Fri, 19 Mar 2010 16:24:43 +0800

BCM1250 has two sibyte core (SB1). The SB1 has four performance counter pairs
(control register and counter register). As a different extention the counter
register of SB1 is 64bit and the bit 40 is the overflow bit. It is bit 31 in
standard mips core (mips32r1/mips32r2/mips64r1/mips64r2).

SB1 is compatible with mips64r1. So the Count/Compare interrupt and the
performace counter overflow interrupt use the same interrupt slot (CAUSE_IP7).

The mainline kernel use SB1's internal timer as its clock event not the
Count/Compare. By default the CAUSE_IP7 is masked in cp0_status.

So SB1 need a different overflow interrupt handler and need to unmask the
CAUSE_IP7.

Signed-off-by: Jack Tan <jack.tan@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>